Fixed namespace for bicep extension methods
