+++ playground/bicep/BicepSample.AppHost/Program.cs
@@ -1,5 +1,3 @@
-using Aspire.Hosting.Azure;
-
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
